Add per-cpu support for map lookup
Per-CPU maps are a bit different, in the sense that they need to return
all the values for each processor core from the lookup.

There is no need to fear breakage with previous versions, as previsouly
lookup on a per-cpu map would fail due to an invalid buffer size.  We
now return a list of all the values for a per-cpu map, instead of the
value directly.